Securex working

+++ src/main/java/byAJ/Securex/controllers/BookController.java
@@ -34,14 +34,15 @@ public class BookController {
         return "redirect:/books/list";
     }
     @RequestMapping("/edit/{id}")
-    public String editBook(@PathVariable("id")int bookid, Model model){
+    public String editBook(@PathVariable("id")long bookid, Model model){
         Book book = new Book();
         book = bookRepository.findOne(bookid);
+
         model.addAttribute("book", book);
         return "bookform";
     }
     @RequestMapping("/delete/{id}")
-    public String deleteBook(@PathVariable("id")int bookid){
+    public String deleteBook(@PathVariable("id")long bookid){
         bookRepository.delete(bookid);
         return "listbooks";
     }
